Implements a ReadWrite Lock for Java DS Joysticks
The locking currently in the DriverStation is locking on "this", which
is not a good idea. This changes to use a specific lock for joystick
data, and changes that to be a ReadWrite lock. Once the data is cached,
all threads can safely read from the cache at any time. Then when there
is new data, the DS thread grabs an exclusive write lock, caches the
data, and then lets the readers go again.

One thing to note is that this does not fix the inefficiency issues with
the GetData method (it holds the lock for 5 ms out of every 20ms),
however to fix that would require JNI changes in a later patch. Then it
would match the C++ locking changes added.

Change-Id: Ide2e68c412e286be3d8322df761cbe7dc473848e